move all js to lib/, working form submission
