+++ site/pdxroasters/static/js/src/main.js
@@ -1,7 +1,6 @@
-import * as map from './sections/map'
-import * as list from './sections/list'
-import scroll from './sections/scroll'
-import forms from './sections/forms'
+import * as map from './lib/map'
+import * as list from './lib/list'
+import scroll from './lib/scroll'
 
 map.addRoasters(list.roasters)
 map.locationFound(list.sortDistanceList)
